requirements.txt
# Backend API + workflow
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
openai>=1.0.0

# RAG (Chroma + sentence embeddings for rag.py)
chromadb>=0.5.0
sentence-transformers>=3.0.0
